Challenge4
Build a food ordering system where the user inputs the food name into the food order and then the order gets passed to the appropriate restaurant on a list and the restaurant prints the receipt on stdout. No need to create a UI, a main function that receives parameters is enough.